Move custom errors into app package
