+++ app/errs.go
@@ -1,4 +1,4 @@
-package dsl
+package app
 
 import "errors"
