+++ tests/specs/admin.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/fabiante/persurl/app"
 	"github.com/fabiante/persurl/tests/dsl"
 	"github.com/stretchr/testify/require"
 )
@@ -33,7 +34,7 @@ func TestAdministration(t *testing.T, admin dsl.AdminAPI) {
 				t.Run(fmt.Sprintf("invalid[%d]", i), func(t *testing.T) {
 					err := admin.SavePURL(purl)
 					require.Error(t, err)
-					require.ErrorIs(t, err, dsl.ErrBadRequest)
+					require.ErrorIs(t, err, app.ErrBadRequest)
 				})
 			}
 		})
